requirements.txt
glfw>=1.4.0
cython==0.29.34
imageio==2.19.3
cffi>=1.10
imagehash>=3.4
ipdb==0.13.9
Pillow>=4.0.0
pycparser>=2.17.0
pytest>=3.0.5
scipy>=0.18.0
cached-property==1.3.1
gitdb==4.0.11
GitPython==3.1.40
gtimer
boto3
boto
cloudpickle==1.3.0
joblib
sk-video
torch
torchvision
opencv-python==4.6.0.66
matplotlib==3.5.1
h5py
termcolor==2.1.1
click
numba
numpy-quaternion
absl-py==0.13.0
jupyterlab==3.0.14
tb-nightly==2.11.0a20221013
imageio-ffmpeg==0.4.4
hydra-core==1.1.0
hydra-submitit-launcher==1.1.5
pandas==1.5.2
yapf==0.31.0
wandb
pylint==2.15.9
six
plantcv
scikit-image
urdfpy==0.0.22
PyOpenGl-accelerate==3.1.7
open3d
pybullet-svl>=3.1.6.4
lockfile
nvisii
gymnasium
colorlog
gym==0.21.0
git+https://github.com/fishbotics/atob@v0.0.1
